backend/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
redis==4.6.0
Flask-Cors==3.0.10
python-dotenv==1.0.0
fakeredis==2.19.0
pytest==7.4.0